+++ packages/progress-bubbles/src/utils.ts
@@ -41,7 +41,8 @@ export function useDimensions<T extends HTMLElement>({ defaults }: UseDimensions
 		observer.observe(node);
 
 		return () => observer.disconnect();
-	}, [node]);
+		// biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
+	}, [node, measureDimensions]);
 
 	return {
 		ref,