+++ packages/vars/stories/Story.css.ts
@@ -8,8 +8,8 @@ export const container = style({
 	backgroundColor: vars.color.bg.ui.secondary,
 	color: vars.color.text.ui.secondary,
 	padding: vars.spacing[500],
-	border: `${vars.border.width[300]} solid ${vars.color.border.ui.primary}`,
-	fontSize: vars.font.size[300],
+	border: `${vars.borderWidth[300]} solid ${vars.color.border.ui.primary}`,
+	fontSize: vars.fontSize[300],
 });
 
 export const local = style({